action review comment

+++ src/main/app/idam/idamClient.ts
@@ -73,8 +73,12 @@ export class IdamClient {
       })
   }
 
-  static invalidateSession (code: string): Promise<void> {
-    const url = `${config.get('idam.api.url')}/session/${code}`
+  static invalidateSession (jwt: string): Promise<void> {
+    if (!jwt) {
+      return Promise.reject(new Error('JWT is required'))
+    }
+
+    const url = `${config.get('idam.api.url')}/session/${jwt}`
     return request.delete(url)
   }
 }
